Add list of all keyboard shortcuts
Summary:
Add a `?` shortcut and button in settings to open a modal listing all of the configured keyboard shortcuts.

Of course, this list is constant right now, but you could imagine making this remappable.

Having a list makes it more discoverable that there are keyboard shortcuts, and generally more accessible.

Reviewed By: quark-zju

Differential Revision: D51326896

fbshipit-source-id: c954afb6052ace16d33b28dfd0ea1f5f8d8b8772