minibench: add an example about filesystem IO
Summary:
I originally wrote this in `P465175275` as part of code review of D31973093 and
D32013158 to raise attention to `buffer_unordered` overhead. It seems this can
be generally useful to reason about raw fs performance (and choose between
async vs non-async for local fs operations). So let's check it in.

Note this is not a universal benchmark that considers details like fs can be
warm or cold or network that helps you decide whether to use async Rust. This
is more for people who know what are tokio tasks and threads, and `tokio::fs`
is just `spawn_blocking` wrappers, and are curious about the raw overhead of
the tokio library features like `buffer_unordered`, for people who want to
kill pointless `Arc` clones.

If you are trying to decide whether to use async Rust from a performance
perspective, write your own benchmarks. Be aware that aside performance, there
are other considerations regarding on async Rust, such as references with
lifetimes are generally harder to use, recursion is harder, stack trace has
more noise, integration with Python (with GIL) needs extra care, etc. See also
the "Why Async?" section from the official async book [1].

[1]: https://rust-lang.github.io/async-book/01_getting_started/02_why_async.html

An example run:

  % buck2 run @//mode/opt :fs_read
  tasks.join_all std::fs::read                       47.413 ms
  tasks.join_all tokio::fs::read                     72.079 ms
  stream.buffer_unordered(1) std::fs::read          847.488 ms
  stream.buffer_unordered(2) std::fs::read          180.042 ms
  stream.buffer_unordered(5) std::fs::read           73.174 ms
  stream.buffer_unordered(10) std::fs::read          44.466 ms
  stream.buffer_unordered(20) std::fs::read          48.515 ms
  stream.buffer_unordered(1) tokio::fs::read        842.601 ms
  stream.buffer_unordered(2) tokio::fs::read        237.096 ms
  stream.buffer_unordered(5) tokio::fs::read         84.668 ms
  stream.buffer_unordered(10) tokio::fs::read        67.564 ms
  stream.buffer_unordered(20) tokio::fs::read        64.711 ms
  1 threaded std::fs::read via channel               79.220 ms
  2 threaded std::fs::read via channel               46.931 ms
  5 threaded std::fs::read via channel               22.941 ms
  10 threaded std::fs::read via channel              14.671 ms
  20 threaded std::fs::read via channel              12.831 ms

Reviewed By: muirdm

Differential Revision: D50986477

fbshipit-source-id: 3112fb4f3fcdb667bb36d0bac2622b01425d3818