Rewrite editedCommitMessages state handling
Summary:
This is a big refactor of the implementation of edited commit messages.

The big idea is that we're simplifying editedCommitMessages to not have a weird default value that depends on the latest commit message. This should fix a very subtle set of bugs that can cause commits to appear duplicated and for diff numbers to be lost.

Edited commit messages are the basis of the sidebar, which allows writing messages and using those messages for operations like amend.

## Before
- editedCommitMessages was an atom family, keyed by hash or `'head'` for writing a new commit.
- editedCommitMessages stored the entire commit message. If nothing was being edited, it would pull from the latest commit message via a default selector.
- fieldsBeingEdited was a selector that stored which fields were being edited. This was related to editedCommitMessages but stored separately.

## After
- editedCommitMessages is still an atom family, but now it has no default state (just `{}`).
- editedCommitMessages no longer stores the entire commit message. Only fields that are edited are stored. So `{}` means no fields are edited. `{Title: 'hello'}` means the title was changed to "hello".
- This means we can derive fieldsBeingEdited from editedCommitMessages
- to start editing a field, just set that field to whatever should be in the edited by default. Usually, this is the underlying latest commit message.

## Why?
By having editedCommitMessage have a default value which uses the latest message, there's potential for a race where we store an old edited value. Once there's a value stored in the atom, it no longer uses its default value selector. That means an old edited value will prevent us from loading the latest value, which means we're liable to amend the old message. This can have several nasty consequences, such as losing the diff number and creating extra diffs.

Plus, I think the code is actually simpler this way.

I introduce a few utilities to make it easier to interact with the state.

This diff actually obviates the need for `type: optimistic` to be part of the edited commit message. I don't remove that here though to keep things as simple as possible. Unfortunately there's still a lot of moving parts here.

Reviewed By: quark-zju

Differential Revision: D51382282

fbshipit-source-id: 21c83472955d5a9ff2f92c7963f8bb92a19ee5c2