goto: set EdenFS parents when changing parents in workingcopy
Summary:
EdenFS contains its own metadata about what its parents are. `workingcopy::set_parents` was originally designed to only alter the treestate of the workingcopy and there is some additional logic that sets up the parents for EdenFS somewhere else, but it's simpler if we git rid of that adhoc logic and get everything in the same place.

Used in later commits.

Reviewed By: muirdm

Differential Revision: D51225916

fbshipit-source-id: 2df01ca5e654ff557981a7398e16ef9518ff312a