test-log-exthook: remove the test and feature
Summary:
The test uses rot13 encoding that does not work out of box on Python 3. The
tested feature was added by "log: add an extension hook-point in
changeset_printer". We don't use such "exthook" to customize
"changeset_printer" and probably don't want to expose the unmodified "ctx" to
random 3rd-party code. So let's remove the test and the feature.

Reviewed By: sggutier

Differential Revision: D51411745

fbshipit-source-id: 779adb5135cc32292d0b48a63cb10f4bdb4d0504