cross-repo sync: allow different names for common push rebase bookmarks between large and small repo
Summary:
## Status quo

In commit sync, there are currently two kinds of bookmaks:
* common pushrebase bookmarks (typically `["master"]`) and
* other bookmarks

Commits pushed to the common pushrebase bookmarks on the small repo get synced to the same bookmark on the large repo.

Other commits get prefixed with `bookmark_prefix` which is configured individually by small repo and pushed.

For instance, assuming `www` is source of truth and forward sync is set up. If the prefix for `www` is `www/`, a push to bookmark `foo` in `www` would result in a push to bookmark `www/foo` in `fbsource`.

This state of configurations assumes that for a given large repo (say `fbsource`), all small repos have the same common pushrebase bookmarks (say `["master"]`).

## Change

As we import git repositories into Mononoke and setup cross-repo-sync between their Mononoke version and fbsource, we will encounter a situation we hadn't encountered before:

The main bookmark on an imported git repo will be named `heads/master` to avoid any conflict with a tag named master which would result in a bookmark named `tags/master`.
We will want to sync such bookmarks to `master` in `fbsource`, not `heads/master`.

To support this, we need to add a per-small-repo config allowing to map from the name of a common pushrebase bookmark in the large repo to the name of the same bookmark in the small repo.

This config is optional. If it is not provided, the mapping should be the identity, keeping backward compatibility with the status quo.

Reviewed By: mitrandir77

Differential Revision: D51391821

fbshipit-source-id: 35413952f4219225746f3ea9a87dbb6cabb90f6f