cross-repo sync: document the pushrebase direction with type-safety
Summary:
In the case of pushrebase, the bookmark is always associated with the target repo.

One way to prove this to ourselves it to look at the implementation of `unsafe_sync_commit_pushrebase_impl`, at the bottom of the call stack (before our change):
https://www.internalfb.com/code/fbsource/[9ba5097d7d6bf5a2e7c1fd9221c4d9ee85759048]/fbcode/eden/mononoke/commit_rewriting/cross_repo_sync/src/lib.rs?lines=1256-1259

In this change, we propagate this knowledge up the call stack through essentially type annotations.

This makes the code harder to misuse for caller up the call stack that may have otherwise lost some of this context.

Reviewed By: mitrandir77

Differential Revision: D51398163

fbshipit-source-id: 558e92d227c0ec3a871ab83d84d4232558769af7