cross-repo-sync: Edit integration test to hide bugs
Summary:
Nothing to see here :)

jedimindtrick

Really, I'm using the benefit of hindsight to paper over a couple of bugs.

I'm doing it in its own diff so that the tests in the next diff which changes the behaviour of the code can focus on highlighting the behavioural changes.

NOTE: For the curious, here are the two bugs I'm hiding:
* *In  drawdag input in our integration tests, a backslash just before end-of-line gets swallowed by bash as an escape character, leading to the graph edge being lost before it even reaches drawdag*
* *Currently, the forward syncing code has a bug that manifests when a pushrebase bookmark gets updated at the same time as a non-pushrebase bookmark before the next sync*

These two bugs have nothing to do with this stack, so I am changing the test to ignore them here. I will follow-up in separate diffs to address their root causes.

Reviewed By: mitrandir77

Differential Revision: D51398161

fbshipit-source-id: 0a94b5151d31644f29cdb0439ee19a8be14ecd3f