clippy: Ignore lint in commit-rewriting
Summary:
This is one case where clippy is getting it wrong:
It sees the
```
let err = Err(...);
err.unwrap_err()
```
and thinks that we may as well omit the `Err` and `unwrap_err()`.

In reality, we pass `err` to a function that wants a `Result` in the meantime, and since `Err` took ownership of the error, we would need to clone to avoid this lint, which wouldn't be worth it.

Respectfully disagree with clippy.

Reviewed By: mitrandir77

Differential Revision: D51398162

fbshipit-source-id: 12ce4425b4671f804b57f9e209917076e0b79a57