Allow empty commits on initial import
Summary:
## What
Allow empty commits to be synced from small to large repo when running the initial-import command in the forward syncer.

## Why?
Because unlike before, where we copied unaltered bonsais from the small to large repo for the initial import, we now have to strip any changes made to git submodules from the bonsais.

This opens up the possibility that a commit becomes empty (i..e no file changes) when synced to the large repo, but we still want those commits to be synced because (a) it has metadata that shouldn't be discarded (e.g. title, author, summary) and (b) it's simpler and less error-prone that trying to "stitch" commit hierarchy/parentage like we did with gitexport (where this was actually needed).

## How
I synced offline with pierrec and mitrandir77 about this, but here is some context:
I spent some time trying to find a way to support this and simultaneously improve the `cross_repo_sync/lib` module. However, this module is very convoluted... There's a lot of duplication and it's very hard to make changes with confidence that we're actually improving and not breaking any random but specific edge case...

So I went with a solution that's not that great, but will work and has some advantages:

**Check the `CommitSyncContext` and enable empty commits only for the initial-import command**.

For posterity, here's my take on the tradeoffs:

**Pros:**
- This approach gives more confidence that all existing use cases won't be affected, because `CommitSyncContext::ForwardSyncerInitialImport` was just defined and is used only in this command.
- It's simple and didn't take a long time to implement.

**Cons:**
- Inserts business logic dependency on `CommitSyncContext`, which, based on my understanding, was introduced mostly for logging/reporting.
- Adds more logic fragmentation to the `cross_repo_sync/lib` module.
  - TBF though it seems to be very hard not to do that, given the current state.

Given these considerations, the time I already spent trying to implement something better and the fact that it's not unlikely we'll invest time in improving the mega repo codebase next half, I thought it would be worth going with this approach.

Please LMK if I missed any relevant point!

Reviewed By: mitrandir77

Differential Revision: D51426083

fbshipit-source-id: 649e9909ec80d86d27303e796749adcd5dadc14a