goto: add a rust helper function for resetting the treestate
Summary:
On EdenFS, when running `goto` the treestate gets reset. On Python we simply would dump the treestate and get a new one, but this is something that we would rather avoid in Rust land as we share the treestate on a number of places.

This moves the implementation in the bindings for resetting the treestate and keeping the same object, but with effectively new data.

Reviewed By: zzl0

Differential Revision: D51311029

fbshipit-source-id: 3af1e98744a503ed407a35bf77229e33aa07f679