Support deriving multiple data types in one command
Summary:
I want to increase the confidence of the tests from the new `initial-import` command in the forward syncer.
One idea I had was to derive all the possible derived data types to make sure there are no hiccups with derivation.

This diff adds support to derive multiple types with one command and the next one adds a flag to derive all the enabled ones (from the repo config).

Differential Revision: D51464164

fbshipit-source-id: c8ab7b45a601f94e74ea51f938894860d5a59438