templatekw: support hg-git/Mononoke committer info formats
Summary:
The hg-git extension uses a different format than Sapling for storing committer info.  In particular, it uses a single extra (also named `committer`) that contains both the committer and committer-date in a single value.  Mononoke uses the same format as hg-git for commits converted from Git.

To support rendering the committer date for both of these, we fall back to the hg-git format if a `committer_date` extra is not present, but a `committer` extra is.

To allow this to be optionally used in smartlog, we update the `sl_date` template alias to use `sl_date_timestamp` as the source of the date.  With this, we can render smartlog with committer dates using `hg sl --config templatealias.sl_date_timestamp=committerdate`.

Reviewed By: zzl0

Differential Revision: D51395319

fbshipit-source-id: ce842aa35d59ae79f494df987bd5a9a267c01b1a