copytrace: add a fast path when similarity percentage is 0
Summary:
When the similarity percentage is set to 0, skip the actual similarity check.

Technically we might still want to check against the "lines" and max cost
settings, but if we think "max cost" as an "optimization" and the main
user facing setting is "similarity" then the fast path seems reasonable.

Reviewed By: zzl0

Differential Revision: D51443692

fbshipit-source-id: 926633c39737e250889c5006b9c2178a07c49b33