status: fix copy-only changes on edenfs
Summary:
If you use "sl cp foo bar --after --force" you end up with "bar" whose only change is addition of copy metadata. When writing the eden dirstate, we were only writing the copy record (i.e. we were not writing the actual record for the file). This incomplete dirstate confused us later. Fix by writing the full record for this file in this case.

The "real" fix is to migrate to the treestate for edenfs, but that is a significant amount of work.

Reviewed By: quark-zju

Differential Revision: D51505484

fbshipit-source-id: 36c780a724bb33c72ae0e51e05f1338d3817fb25