Fix UX hurdle when using `--local-configerator-path`
Summary:
With these default settings being incorrect, it wasn't possible to use `--local-configerator-path` to a mononoke_app without also passing `--tunables-local-path` and `--config-tier`.
This was a minor headache every time.

In the test plan, we demonstrate with gitimport, but this behaviour would apply to any of our CLIs implemented with the `mononoke_app` framework.

Reviewed By: mzr

Differential Revision: D51549531

fbshipit-source-id: 2cd1aef25bf82fef9a0cb35e995bb06fc9ae64ec