Add an option to discard git submodules on import
Summary:
In general, we want the existing default behaviour, which is to represent the submodules in our data-model. However, in some cases, it may be OK to discard submodules and it may be the desired behaviour.

For instance, for an import of a git repo that was already manually synced to Mononoke, where we only want to import the old commits for a best effort view of the history of the repo, it may be acceptable to be lossy.

Reviewed By: gustavoavena

Differential Revision: D51551760

fbshipit-source-id: 206d2fa44e67a431157526da901b5f366ad9985a