commit_graph: make segment split deterministic
Summary:
Switch from `HashMap/Set` to `BTreeMap/Set` in the implementation of `ancestors_difference_segments` so that the segments it produces are deterministic.  This is not strictly necessary, but makes writing tests easier.

The existing test for commit graph segments can now be updated to check the output, since it is deterministic.

Add a test for the corresponding edenapi endpoint to demonstrate that it works, however this reveals a problem: we may include references to nodes that aren't described in the response.

Differential Revision: D51469897

fbshipit-source-id: a597ea12986ae4ff9d8f35edd63af330cb3f848d