commit_graph: ensure returned segments include valid locations for all parents
Summary: Fix the problem exposed in the test in the previous diff by ensuring that we return valid locations for all segment parents.  To do this, we defer the population of locations for parents that are outside the set we are considering in `disjoint_segments`.  Each call to `disjoint_segments` now returns the locations of any commits that could potentially be parents, even if they are in the middle of a segment.  We then use these to fill in all of the segment parents that are part of the set of commits that are being returned.

Reviewed By: YousefSalama

Differential Revision: D51469896

fbshipit-source-id: ee3fc4df3da3c38affc66e40d186e46e7b5bed61