Change prefetch profile generation to allow for no compression
Summary:
We want the prefetch profiles generated for RL to not be converted into globs. We see a large perf improvement went utilizing just filenames instead of globs. This change allows for the creation of a unchanged prefetch profile.

The changes will remove the chosen prefetch profiles from the original chronos job. I will create a new chronos job that will be daily and run the --no-compression flag.

I moved the glob_filenames_blocklist config to the sandcastle gen command. This should be better as we'd only need to call the config in one spot.

Reviewed By: genevievehelsel

Differential Revision: D51532448

fbshipit-source-id: a48fb0ec127e0b3f12e1324b68ea09cf38b2aa98