service: fix a misuse of `ImmediateFuture::ensure`
Summary:
The return value of the passed in lambda is dropped in the `ensure` method,
thus the future returned by `completeInvalidations` would never be waited on,
potentially leading to weird bugs.

This is only used in a debug CLI command so this shouldn't have any effect, but
let's add a static_assert check in the ImmediateFuture code to make sure it
isn't misused elsewhere or in the future.

Reviewed By: genevievehelsel

Differential Revision: D51254287

fbshipit-source-id: 9fcfabbf2318aef739ca44daf1543dc0b83f82e3