inode: revamp blob loading
Summary:
The blob loading code used to rely on an unsafe detached future that could lead
to deadlocks. To fix this, the code needs to be reworked quite significantly
to allow the loading future to be chained and added to a proper executor by the
calling context.

Simply chaining the future is unfortunately not sufficient: it prevents the
loading context from being interrupted by a truncation, this can be solved by
using a `folly::collect` with both the promise and the load future. In the case
where the load is interrupted, the promise will be set and the collect will
early complete. Since `folly::collect` is only early completing when one of the
future fails, some boilerplate code had to be added to handle the
`BrokenPromise` case.

However the solution above exposes another very subtle issue: if the returned
future is dropped, the loading will never complete but the inode will stay in
its loading state. Thus future loads will hang waiting for a promise that will
never be set. Solving this can be done by detecting when the loading future is
dropped and simply resetting the loading state.

Reviewed By: jdelliot

Differential Revision: D50415722

fbshipit-source-id: 33833b744ca0bcacfc7bc8c53a463dd99a327889