pull: implement pull using new commit graph segments endpoint
Summary:
Similar to what we did for clone, we implement fast-pull using the commit graph segments endpoint by converting the response to pull data.

There is some additional complexity here, as pull data can refer to commits that are outside of the set of commits in the pull data.  In fact, a segment parent could be a distant ancestor of one of the common heads (consider, for example, the merge of an ancient long-lived branch).  To handle this, we assign all external parents their own temporary ids.  These are just used so that we can have an entry in the idmap, and are not part of the segments themselves.  When the pull data is imported into the local dag, all of the ids are either matched up with existing commits, or assigned new ids in the newly imported segments.

Reviewed By: quark-zju

Differential Revision: D51529308

fbshipit-source-id: a1f97c8e757bdc9d9f66b62614666348faf69be2