thread fault injector to HgDatapackStore
Summary:
We had a crash in the HgDatapackStore because we held a reference into
an object in the eden config. I fixed it in D51133312, but I had punted on
writing a test since it needed some plumbing.

The first piece of that plumbing is being able to inject an artifical delay
from when the reference into the EdenConfig is aquired and used.

FaultInjector can help us do that. Plumbing that down into the hgDatapackStore
in this diff.

Reviewed By: genevievehelsel

Differential Revision: D51621073

fbshipit-source-id: fbdd29d8bebadec5ca932e3bc7e9b6c8c2654352