split out TestConfigSource
Summary:
We had a crash in the hgDatapackStore because we held a reference into
an object in the eden config. I fixed it in D51133312, but I had punted on
writing a test since it needed some plumbing.

The next piece I need is to be able to trigger an EdenConfig reload. The
inode test stuff already has TestConfigSource that allows doing that. Pulling
this code out of cpp files so I can reuse it in my test.

Reviewed By: genevievehelsel

Differential Revision: D51621102

fbshipit-source-id: 2f07325b2c476d8c188af977455166709dcc9619