test for HgDatapackStore/EdenConfig crash
Summary:
We had a crash in the hgDatapackStore because we held a reference into
an object in the eden config. I fixed it in D51133312, but I had punted on
writing a test since it needed some plumbing.

We finally have all the plumbing we need so now we can write a test that
causes the edenconfig to be reloaded while a reference is being held to a
config value.

Reviewed By: genevievehelsel

Differential Revision: D51621134

fbshipit-source-id: 5b7afd0e7971b58adb3ed6e51c8895f0d95275f6