edenfs_config_manager: fix corp gk checking
Summary:
When testing this initally on macOS, I was using `buck run eden/fs/config/facebook:edenfs_config_manager`, which runs as the correct user, but when this will be ran by systemd timers, the current user would be `root`. This changes the gk checker to check for the configured user instead

The behavior on Windows should be the same because the config manager already just runs as the user. I will retest it anyway though

Reviewed By: kmancini

Differential Revision: D51437085

fbshipit-source-id: b9a454c3210ace4e48e5580f5fd504448f56d620