remove digest status tracker from CASClient::co_download* return type
Summary:
FWICT, co_download* method return DigestStatusTracker only when the method doesn't throw an exception. And  this method throws an exception whenever status of any digest is not OK (i.e. we failed to download it). Ergo, the only result we can get in digest status tracker is OK. Ergo, we don't really even need to return DigestStatusTracker in the first place.

This is what this diff does - removes DigestStatusTracker from DownloadResult.

Differential Revision: D51558546

fbshipit-source-id: 1cc2016fb56d49571c65021dbb2f8a7658c0b475