test_manifest: Introduce TestManifest
Summary:
Context: I need two test manifest types, one (TestShardedManifest) should be a sharded version of the other (TestManifest). To be used in integration tests showcasing an optimized workflow for backfilling derived data types that have the same context as existing types but are only structured differently. Relying on production types would be inconvenient as we would need to change the tests whenever we want to modify them.

TestManifest is a manifest type intended only to be used in tests. It contains only the file names and the maximum basename length of all files in each directory.

Reviewed By: gustavoavena

Differential Revision: D51464853

fbshipit-source-id: ef97ba554197e5720ffad39d8e96a466cc1f59e0