bssm_v3: implement derivation logic
Summary:
Implements the derivation logic of BssmV3. The derivation works by taking the changes in the bonsai changeset and transforming the paths by prepending each with the reverse of the basename (this is defined in BssmPath in the original Bssm code). Then using the generic manifest derivation code to derive the new manifest out of the transformed changes.

Note: There is a bug in the original Bssm derivation code due to mishandling of implicit deletes, implicit deletes must be expanded before doing the bssm transform on the changes. This is because paths that were prefixes of each other before the transform can become independent afterwards, and the only implicit deletes after the transform that are perserved are those that share the same basename. For example:

Before bssm transform:
- Parent manifest has one file: a/b/c
- Bonsai changeset adds a file: a/b
- a/b/c is implicitly deleted.

After bssm transform:
- Parent manifest has one file: c/a/b/c
- Bonsai changeset adds a file: b/a/b
- no implicit deletes.

The added unit test catches this bug.

Reviewed By: markbt

Differential Revision: D51489411

fbshipit-source-id: 329c1d6173875da1471197e2476ce50c10c96df5