test_sharded_manifest: derived data boilerplate
Summary:
Context: I need two test manifest types, one (TestShardedManifest) should be a sharded version of the other (TestManifest). To be used in integration tests showcasing an optimized workflow for backfilling derived data types that have the same context as existing types but are only structured differently. Relying on production types would be inconvenient as we would need to change the tests whenever we want to modify them.

This diff adds the boilerplate required when adding a new manifest derived data type.

Reviewed By: markbt

Differential Revision: D51465891

fbshipit-source-id: 41519638b0c9e79a782ae1fc7030119b1e2aa520