test_manifest: implement derivation logic
Summary:
Context: I need two test manifest types, one (TestShardedManifest) should be a sharded version of the other (TestManifest). To be used in integration tests showcasing an optimized workflow for backfilling derived data types that have the same context as existing types but are only structured differently. Relying on production types would be inconvenient as we would need to change the tests whenever we want to modify them.

This diff implements the actual derivation logic of TestManifest.

Reviewed By: markbt

Differential Revision: D51464854

fbshipit-source-id: b5e1a94aa21d53013721ba8606f6eb63c4dffb76