derived_data_manager: implement derive_from_predecessor
Summary:
Extends the derived data manager with a method that derives data for a changeset using other derived data types without requiring data to be derived for the parents of the changeset.

Will be used to parallelize backfilling derived data by slicing the commits of a repository, deriving data for the boundaries of the slices using this method, and then deriving data for the rest of the commits for all slices in parallel using the normal derivation path.

Differential Revision: D51467401

fbshipit-source-id: 5d6efd645ed92bf536c04d76761b5727297f66c0