commit_graph: implement process_topologically
Summary:
Adds a method to the commit graph that run a given `process` closure on all of the given changesets in local topological order, running as many of them concurrently as possible.

Note: local topological order here means that all parents of a changeset that are contained in the input `cs_ids` are processed before itself, but if two changesets are ancestors of each other and some of the changesets in the path betwen them are not given in `cs_ids`, they are not guaranteed to be processed in topological order.

Reviewed By: markbt

Differential Revision: D51487587

fbshipit-source-id: 4bb20dce51d744191fb838eabde0f377d0288586