test_sharded_manifest: introduce TestShardedManifest
Summary:
Context: I need two test manifest types, one (TestShardedManifest) should be a sharded version of the other (TestManifest). To be used in integration tests showcasing an optimized workflow for backfilling derived data types that have the same context as existing types but are only structured differently. Relying on production types would be inconvenient as we would need to change the tests whenever we want to modify them.

TestShardedManifest is a sharded version of TestManifest intended only to be used in tests. It contains only the file names and the maximum basename length of all files in each directory.

Reviewed By: markbt

Differential Revision: D51465244

fbshipit-source-id: f02b74ff581f704dfe07ff40b796b3141a48b363