derived_data: add derive_from_predecessor to BonsaiDerivable trait
Summary:
Extends the BonsaiDerivable trait with a method that derives data for a changeset using other derived data types without requiring data to be derived for the parents of the changeset.

Will be used to parallelize backfilling derived data by slicing the commits of a repository, deriving data for the boundaries of the slices using this method, and then deriving data for the rest of the commits for all slices in parallel using the normal derivation path.

Reviewed By: markbt

Differential Revision: D51467068

fbshipit-source-id: bcc311a35edc31753dba5d2242bb3d137e47397a