newadmin: add derived-data slice subcommand
Summary:
Adds a subcommand `mononoke_newadmin derived-data slice` that uses the slice_ancestors method of the commit graph to slice the ancestors of the given commits using their generation numbers. Each slice is defined using a list of changeset ids `slice_frontier` and a minimum generation `slice_start`, where all changesets that are ancestors of any of `slice_frontier` and have a generation number higher than or equal to `slice_start` are considered to be part of the slice.

If the rederive flag is provided all ancestors of the given changesets will be sliced, otherwise only the ancestors that aren't derived yet will be sliced.

Also extends ChangesetArgs with a flag `all_bookmarks` to allow specifying all bookmarks at once.

Differential Revision: D51482456

fbshipit-source-id: 7df1366f8c2002dd339781ec0ee57771ea93cded