bssm_v3: implement derive_from_predecessor using skeleton manifests
Summary:
Implements derive_from_predecessor for BssmV3 to be able to derive BssmV3 from a skeleton manifest without requiring BssmV3 to be derived for the parents first.

Since skeleton manifest and BssmV3 have different structures we can't make use of the generic manifest derive_from_predecessor code. Instead we can get a stream over the leaf entries of the skeleton manifest, chunk it into a reasonable size (I picked 10000), transfrom all the changes in the chunk using the bssm transform code, then update BssmV3 with them. For updating BssmV3 we reuse BssmV3 bonsai derivation code, since all the leaf entries of skeleton are independent (none of them are a prefix of another) they will remain so after the bssm transform, so the derivation code becomes exactly the same as adding all the entries into the manifest.

Differential Revision: D51489414

fbshipit-source-id: 17af1d4c15fc5b1fab1286a7eacd91a100d14740