indexedlog: attempt to handle SIGBUS from btrfs checksum mismatch
Summary:
We have seen edenfs crashing with SIGBUS in production. They are caused by
btrfs checksum errors when reading mmap buffers.

This diff adds a SIGBUS handler that zero-fills the page causing SIGBUS on
demand as an attempt to avoid SIGBUS crash. The idea is that the zero-filled
page will fail the indexedlog checksum somehow so it will become a (seemingly
regular) error, and might trigger re-tries (ex. re-fetch via network) to
transparently fix the issues without user intervention.

Considerations and discoveries:
- `setjmp` error handling: hard to verify soundness in Rust.
  See https://github.com/rust-lang/rfcs/issues/2625
- Simply `return` from the signal handler: The SIGBUS will happen again and the
  program enters an infinite loop trying to read the bad data.
- `mmap` support to return zero pages without SIGBUS: There were some interests
  and attempts in the Linux kernel community. However, nothing usable is merged
  yet.

This signal handler assumes the SIGBUS behavior stay consistent during the
process lifetime. Namely, for previously returned `Ok` data, if read again
later from the same process, it should remain `Ok` not `SIGBUS`. It's okay to
get `Ok` consistently, and after a hard reboot, only get `SIGBUS`. While
practically this assumption seems okay to make (and the use of `mmap` relies on
the assumption that the returned `Ok` data remain unchanged), it is hard to
predicate all corner cases, and the code is optimized for easy-to-understand
rather than "async signal safety". Therefore, the feature is default off for
now.

Reviewed By: zzl0

Differential Revision: D51623839

fbshipit-source-id: 506b676ee1c6d4075a02417622adb0d3477c7a7a