Don't autofill "Temporary Commit" as default title
Summary: A consequence of the change to fix empty titles when committing meant that if you don't have a commit template, your title would default to "Temporary Commit" in the commit form. This was not intended, and is a little annoying. I'd rather have an empty title field that I can edit as I please.

Reviewed By: quark-zju

Differential Revision: D51676830

fbshipit-source-id: 186cba58f88677285dd516ecebfdab1fceb7f309