Use edited message when running fold preview
Summary:
After selecting and clicking "combine commits", you see a preview of the fold as a fake commit in the tree. This commit is auto-selected, and in the commit sidebar you can see and edit the combined commit message as you like. Then, when clicking "run combine", it should use your latest edits instead of only using the merged message.

This required shuffling some things around. We need to modify the FoldOperation, but it's read-only inside of the recoil state. So we make a new copy of the fold operation with an updated message by taking the edited commit message for this fold range. We compute the latest message by taking the edited message and merging it with the already merged message.

Reviewed By: quark-zju

Differential Revision: D51641870

fbshipit-source-id: f15e22cabdb595f0962776a6383e963a12839387