Separate fake commits for preview fold and optimistic fold
Summary:
It was a bit janky when running a previewed fold operation, since it would leave your fields edited, since the optimistic commit was the same as the preview commit.

instead, we can have separate fake commit for the preview commit and the optimistic commit. Now, running the operation shows the optimistic commit, so fields aren't being edited. This make it feel like it's running the command.

Reviewed By: quark-zju

Differential Revision: D51641879

fbshipit-source-id: cf16373151feedaa3627862daaceca9dd835acde