Add "Select all commits" button
Summary:
Add a button to the bulk actions menu to select all commits. This gives quick access to acting in bulk on a granular selection from the sidebar.

We'll need to add more bulk actions to the sidebar, for example rebasing your selection onto a specific commit.

We can also add a shortcut for this selection which will make this even more useful.

Reviewed By: quark-zju

Differential Revision: D51686535

fbshipit-source-id: f8d05986844e58c2909fc0dcf9786f593219f0d4