Bulk rebase selection
Summary:
Add an operation that bulk rebases a set of commits, and add a button to use it to the multi-select sidebar.

This means you can select many commits and then bulk rebase them.

We use `rebase --rev A --rev B ...` so you can rebase multiple things at once, and it even still uses successors properly for handling optimistic state.

Reviewed By: quark-zju

Differential Revision: D51688168

fbshipit-source-id: e5b5c2b41533c0312592b03b12171a47c3572f55