Add bulk actions dropdown
Summary:
Add a new dropdown for actions which affect all your commits/diffs.

This will include rebasing all draft commits, cleaning up all commits, resubmitting all out of sync commits, etc.

I think a new dropdown is justified for this, though each new dropdown makes the toolbar more intimidating and confusing, so we do need to be careful with it. Definitely better in this dropdown than just as a "rebase all" button in the toolbar directly (like the old ISL).

Reviewed By: quark-zju

Differential Revision: D51683871

fbshipit-source-id: 74950f3bb87b496504738c8b9c3f085fe042a9c3