Bulk rebase
Summary:
Add in actual support for rebasing all draft commits onto a destination.

This makes it easy when you have lots of small commits that you want to rebase.

We can actually mostly reuse the Suggested Rebase button for this, except we need a new Operation. The new Operation is mostly like rebase, except it knows to act on all commits.

We just use a simple revset so we don't have to manually list commits to rebase: `draft() & date(-14)`. The date is a small optimization that only rebases what's visible in ISL, and not all your draft commits. This makes it more predictable in the UI.

We don't need a preview applier, since we don't show a preview of the rebase, we just go for it (this matches what the other suggested rebase button does).

Reviewed By: quark-zju

Differential Revision: D51683872

fbshipit-source-id: 542e2e15dae0caaae22da5b261e7f9eaa9eeee9e