prevent Rust/Python repos from falling out of sync during sparse operations
Summary:
It's possible that commits created in the Python repo are not available when they are requested in the EdenFS Rust repo. This is because the data hasn't been flushed to disk and picked up by the Rust repo. This diff ensures that all this information is flushed to disk and re-read by the Rust repo if any commit lookup fails.

We also run into the same problem with the tree/file stores. We move flushing/invalidation logic before reading those to avoid missing data.

Reviewed By: muirdm

Differential Revision: D51499993

fbshipit-source-id: c92fc2dbdf1bec5ac7fed57331437c8008e223d2