prevent misc. sparse profile logic from running with non-filtered Eden repos
Summary:
There are some places scattered across the codebase that check if the repo has the sparsematch() attribute. With the addition of edensparse, all repo objects will have this attribute. However, we only want to run the sparse logic if

1) the repo is sparse
2) the repo is using edensparse

This diff prevents the misc. sparse/edensparse logic from running if neither of those conditions are true.

Reviewed By: quark-zju

Differential Revision: D50990329

fbshipit-source-id: 7c70469f2f9ce4e51091cb99041950686842696c