prevent sparse logic from running for eden checkouts
Summary: The edensparse and sparse extensions share a lot of logic, including an attribute named sparsematch(). This allows a lot of the existing sparse profile logic to be used for edensparse, but it also causes them to conflict in certain situations. This diff prevents any sparse (not filter) logic from running if an eden checkout is in use.

Reviewed By: quark-zju

Differential Revision: D50990330

fbshipit-source-id: e45003d578f0ca4daa614d7383e9d8e429c73d10