allow in-place eden updates
Summary:
The edensparse extension expects to be able to perform in-place checkouts to apply filters to the working directory. In order to do this, we need to remove the early-return optimization that happens when the checkout destination is the same as the current commit.

In place checkouts are used in the next diff.

Reviewed By: muirdm

Differential Revision: D51325072

fbshipit-source-id: 149e6d3ff765a8e0381e495a19e883099a45b6a3