add helpers for determining/reporting sparse flavor
Summary: The next few diffs will modify the existing sparse extension so that it's compatible with the new `edensparse` extension I'm introducing. Sometimes in the code we need to understand if we're running under a `sparse` or `edensparse` checkout. These helpers report which type of repo we're running as and help provide accurate output to users.

Reviewed By: muirdm

Differential Revision: D51325071

fbshipit-source-id: 5c8551a0a7914cf4c317806ca1b0639db973133a