implement disable command
Summary: Implements the disable command for the edensparse extension. Since only 1 filter can be active at a time, it doesn't make sense to accept a list of profiles to disable. To work around this, I slightly modified the sparse profile code to disable all (aka 1) active profiles if an edensparse repo is in use.

Reviewed By: muirdm

Differential Revision: D51336777

fbshipit-source-id: 3740c797691c0ba86a602948b8966df4dc01da4a